requirements-dashboard.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.7
jinja2==3.1.2